docs(#78,#80,#82): rename test files, fix CLAUDE.md, spec housekeeping
#78 — Rename misaligned test files to match spec numbering:
  - s06_pipeline_inheritance.rs → s07_pipeline_inheritance.rs (spec is §7)
  - s08_http_runner.rs → r05_http_runner.rs (spec is runner r05, not core s08)
  - s15_skills.rs → s06_skills.rs (spec is §6, not §15)
  Update mod.rs declarations.

#80 — Fix --once Flow step 3 in CLAUDE.md: was "resumes via --resume"; corrected
  to "steps run isolated by default; set resume: true on a step to resume."

#82 — Six spec housekeeping fixes:
  - CLAUDE.md: update result_large_err list to reflect all current files; soften
    "deterministic chain" tagline to "declared sequence of automated steps"
  - s03: remove duplicate timeout_seconds; clarify on_error not yet parsed
  - s19: add implementation status note (Three Tiers is vision, not current);
    fix stale claim that per-step runners are "always headless" — they now inherit
    Session.headless (corrected in PR-E)
