+++ ail-core/tests/spec/mod.rs
@@ -1,3 +1,4 @@
+mod r05_http_runner;
 mod s03_file_format;
 mod s04_execution_model;
 mod s04_log_provider;
@@ -5,8 +6,8 @@ mod s05_3_on_result;
 mod s05_5_context_steps;
 mod s05_append_system_prompt;
 mod s05_step_specification;
-mod s06_pipeline_inheritance;
-mod s08_http_runner;
+mod s06_skills;
+mod s07_pipeline_inheritance;
 mod s08_multi_runner;
 mod s08_runner_adapter;
 mod s09_sub_pipeline;
@@ -15,7 +16,6 @@ mod s10_model_config;
 mod s11_template_variables;
 mod s12_step_conditions;
 mod s13_sqlite_provider;
-mod s15_skills;
 mod s17_error_handling;
 mod s18_materialize;
 mod s21_mvp;
