--- ail-core/tests/spec/s06_pipeline_inheritance.rs
@@ -1,13 +1,0 @@
-/// SPEC §6 — extends: field traverses parent pipeline files
-#[test]
-#[ignore = "not yet implemented — SPEC §6"]
-fn extends_field_is_parsed() {
-    todo!()
-}
-
-/// SPEC §6 — materialized chain includes all inherited steps with origins
-#[test]
-#[ignore = "not yet implemented — SPEC §6"]
-fn materialize_traverses_inheritance() {
-    todo!()
-}
