+++ ail-core/tests/spec/s06_skills.rs
@@ -1,0 +1,6 @@
+/// SPEC §6 — skill: steps (planned for v0.2+)
+#[test]
+#[ignore = "not yet implemented — SPEC §6"]
+fn skill_step_loads_and_executes_skill_pipeline() {
+    todo!()
+}
