+++ ail-core/tests/spec/s07_pipeline_inheritance.rs
@@ -1,0 +1,13 @@
+/// SPEC §7 — extends: field traverses parent pipeline files
+#[test]
+#[ignore = "not yet implemented — SPEC §7"]
+fn extends_field_is_parsed() {
+    todo!()
+}
+
+/// SPEC §7 — materialized chain includes all inherited steps with origins
+#[test]
+#[ignore = "not yet implemented — SPEC §7"]
+fn materialize_traverses_inheritance() {
+    todo!()
+}
