--- ail-core/tests/spec/s15_skills.rs
@@ -1,6 +1,0 @@
-/// SPEC §15 — skill: field loads and executes a skill pipeline
-#[test]
-#[ignore = "not yet implemented — SPEC §15"]
-fn skill_step_loads_and_executes_skill_pipeline() {
-    todo!()
-}
